Precision changes, grid point information, and adaptions for extra post-processing.
The grid point value stored in the new files is now the grid centre,
not the lower-left corner. Also northing and easting data arrays
have been added to the output files.

The BNG to lat/lon code has been changed to double precision as
much as possible, to try to make it more accurate. However there
can still be upto ~100m movement in grid points at the extreme
corners of the BNG region, so more work is needed on this.

For post-processing purposes you can also change the names of the
lat / lon dimensions, as xarray doesn't like these to share a name
with existing data arrays.